Kim - I did some crazy stuff with scrolling
worked on the homepage at a crazy time of night.
The session kicked my butt so I decided to do a hackathon :P

Hope you like it!

I added some new CSS and new HTML to demonstrate the scrolling.

I will add the rest of the wireframe design in for the homepage when I get some zzzzzz.